+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Fixed
 - Fixed ssh multiplexing master connection initializing
 - Fixed `dep ssh` command [#1204]
+- Fixed `dep config:current` task
 
 ## v5.0.0-beta.3
 [v5.0.0-beta.2...v5.0.0-beta.3](https://github.com/deployphp/deployer/compare/v5.0.0-beta.2...v5.0.0-beta.3)